add - css : fadout capability
This change is meant to replace the `hide-btn-text `logic. 

With this method, we can choose how much the text have to fade out, with a default of 0.2.

Once this merge get accepted, we can safely remove the `hide-btn-text` and his related `@Input()` logic.
Wish will make the directive even lighter.